Add a cmake config for testing with googletest
This adds the google test sources as a submodule and some
configuration to build them as a module.

To build the test we add a new executable target 'testcmd'
in src/test/ that compiles the source files under src/test/.

We also add a run_tests target that just executes testcmd.