Fix a bug in inexor::util::split_find_last
Before this commit, splits such as split_find_last("aba", "ab")
would return no match. This was the case because find_last_of
also matches partial strings, so it matches the last a of
aba, because the search string ab begins with a.

This is fixed by using string::rfind instead of
string::find_last_of

In creating this fix we also add a custom to_string function
which can work on many types.